multiAgent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
openai==1.3.5
anthropic==0.7.4
asyncpg==0.29.0
sqlalchemy==2.0.23
sqlalchemy[asyncio]
python-dotenv==1.0.0
slowapi==0.1.8
httpx>=0.23.0
aiohttp==3.9.1
pgvector==0.2.4
python-multipart==0.0.6
alembic==1.12.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
